requirements.txt
django_ckeditor==5.7.1
Django==2.2.7
